# perf(serialize): optimized sorting for small objects