--- a/src/serialize.ts
+++ b/src/serialize.ts
@@ -46,11 +46,20 @@ const Serializer = /*@__PURE__*/ (function () {
         return typeA < typeB ? -1 : 1;
       }
 
-      if (typeA === "string" || typeA === "number") {
-        return a < b ? -1 : 1;
+      switch (typeA) {
+        case "object":
+        case "function":
+        case "symbol": {
+          const serializedA = this.serialize(a, true);
+          const serializedB = this.serialize(b, true);
+          if (serializedA === serializedB) {
+            return 0;
+          }
+          return serializedA < serializedB ? -1 : 1;
+        }
       }
 
-      return this.serialize(a, true) < this.serialize(b, true) ? -1 : 1;
+      return a < b ? -1 : 1;
     }
 
     sort<T>(array: T[], compare?: (a: T, b: T) => number): T[] {