For smaller arrays with less than ~10 elements we can use a simple insertion sort.

```sc
 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small 2191ms
     name                       hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ d7f7f9d  3,085,507.01  0.0003  0.2968  0.0003  0.0003  0.0006  0.0007  0.0014  ±0.23%  1542754
   · ohash @ dev      4,215,331.50  0.0002  2.6968  0.0002  0.0002  0.0004  0.0004  0.0006  ±1.43%  2107666   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large 1210ms
     name                    hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ d7f7f9d  14,261.90  0.0641  0.3515  0.0701  0.0687  0.1101  0.1909  0.2658  ±0.48%     7131
   · ohash @ dev      14,788.36  0.0621  0.4958  0.0676  0.0666  0.0998  0.2151  0.2611  ±0.51%     7395   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:256, size:small 1208ms
     name                    hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ d7f7f9d  12,785.12  0.0684  0.6685  0.0782  0.0726  0.2825  0.3235  0.3818  ±1.08%     6393
   · ohash @ dev      16,530.58  0.0531  0.4127  0.0605  0.0568  0.2606  0.3010  0.3378  ±1.01%     8266   fastest

 ✓ test/benchmarks.bench.ts > benchmarks > serialize - presets > count:128, size:large 1208ms
     name                   hz     min     max    mean     p75     p99    p995    p999     rme  samples
   · ohash @ d7f7f9d  1,660.30  0.5209  1.3680  0.6023  0.5744  1.2529  1.3128  1.3680  ±1.74%      831
   · ohash @ dev      1,673.26  0.5201  1.4171  0.5976  0.5679  1.2493  1.2953  1.4171  ±1.75%      837   fastest

 BENCH  Summary

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:small
    1.37x faster than ohash @ d7f7f9d

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:1, size:large
    1.04x faster than ohash @ d7f7f9d

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:256, size:small
    1.29x faster than ohash @ d7f7f9d

  ohash @ dev - test/benchmarks.bench.ts > benchmarks > serialize - presets > count:128, size:large
    1.01x faster than ohash @ d7f7f9d
```
	
```sc	
clk: ~5.23 GHz
cpu: AMD Ryzen 9 7950X 16-Core Processor
runtime: bun 1.2.4 (x64-linux)

benchmark                   avg (min … max) p75 / p99    (min … top 1%)
------------------------------------------- -------------------------------
• count:1, size:small
------------------------------------------- -------------------------------
ohash @ d7f7f9d              206.22 ns/iter 200.34 ns ▅█
                    (189.48 ns … 539.76 ns) 426.14 ns ██
                    (  0.00  b … 329.00  b)   4.34  b ██▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

ohash @ dev                  174.60 ns/iter 164.94 ns █
                    (156.44 ns … 531.82 ns) 397.53 ns ██
                    (  0.00  b … 660.00  b)  11.91  b ██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   1.18x faster than ohash @ d7f7f9d

• count:1, size:large
------------------------------------------- -------------------------------
ohash @ d7f7f9d               72.96 µs/iter  72.69 µs  █▄
                       (60.95 µs … 1.75 ms) 126.71 µs  ██
                    (  0.00  b … 576.00 kb)  11.34 kb ▂███▅▄▃▃▃▂▂▁▂▁▁▁▁▁▁▁▁

ohash @ dev                   71.43 µs/iter  72.83 µs   █
                       (60.21 µs … 1.41 ms) 111.64 µs   █▆
                    (  0.00  b … 528.00 kb)   6.84 kb ▁█████▅▄▃▂▂▂▂▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   1.02x faster than ohash @ d7f7f9d

• count:256, size:small
------------------------------------------- -------------------------------
ohash @ d7f7f9d               66.01 µs/iter  59.63 µs  █
                       (52.27 µs … 1.94 ms) 136.16 µs  █
                    (  0.00  b … 528.00 kb)  17.16 kb ▃█▆▂▃▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

ohash @ dev                   52.22 µs/iter  46.57 µs  █
                       (41.42 µs … 1.73 ms) 114.14 µs  █
                    (  0.00  b … 528.00 kb)  12.80 kb ▂█▃▁▂▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁

summary
  ohash @ dev
   1.26x faster than ohash @ d7f7f9d

• count:128, size:large
------------------------------------------- -------------------------------
ohash @ d7f7f9d              491.75 µs/iter 480.37 µs   █▃
                      (421.59 µs … 2.45 ms) 723.61 µs  ███
                    (  0.00  b … 792.00 kb)  64.01 kb ▃████▅▃▂▂▁▁▁▁▂▃▃▂▂▁▁▁

ohash @ dev                  492.52 µs/iter 482.14 µs   █▆
                      (422.16 µs … 2.36 ms) 723.95 µs  ███▂
                    (  0.00  b …   1.03 mb)  58.92 kb ▃████▄▃▂▂▁▁▁▁▂▃▃▂▂▂▂▁

summary
  ohash @ d7f7f9d
   1x faster than ohash @ dev
```