@pi0 2025-03-25
Do you recall if we have the same trade-off in v8 based runtimes / Node.js specifically?